// app/src/components/users/Users.types.ts
export type User = {
  id: number;
  email: string;
  password: string;
  address: string;
  city: string;
  postalCode: string;
  country: string;
  phone: string;
};

export type UserPostData = {
  email: string;
  password: string;
  address: string;
  city: string;
  postalCode: string;
  country: string;
  phone: string;
};